Move new classes to new namespace
